# ARROW-3320: [C++] Improve float parsing performance