… cpp/src/arrow/symbols.map
… cpp/thirdparty/versions.txt