This relies on Google's double-conversion library at https://github.com/google/double-conversion

Before:
```
---------------------------------------------------------------------
Benchmark                              Time           CPU Iterations
---------------------------------------------------------------------
[...]
BM_FloatParsing<FloatType>          4621 ns       4620 ns     151802   1.65125M items/s
BM_FloatParsing<DoubleType>         4629 ns       4628 ns     150171   1.64846M items/s
```

After:
```
---------------------------------------------------------------------
Benchmark                              Time           CPU Iterations
---------------------------------------------------------------------
[...]
BM_FloatParsing<FloatType>           563 ns        563 ns    1240915   13.5616M items/s
BM_FloatParsing<DoubleType>          313 ns        313 ns    2227610   24.3934M items/s
```