@wesm 2018-09-25
Is google/double-conversion a small enough codebase to vendor similar to what you did with Abseil (either submodule or simply `cp -r && git add`?