@wesm 2018-09-25
I see this is in conda-forge so it may not be a big deal