@pitrou 2018-09-25
The codebase looks small indeed. The `.git` directory is 7.2MB and the source tree itself is 19M.
The `conda-forge` package is apparently not built for Windows, and it only provides the shared library.

OTOH, building double-conversion from scratch (as an external project) is reasonably quick too. What do you suggest?