@wesm 2018-09-25
Let's stick with ExternalProject for now; this perhaps puts more pressure for the upstream project and conda-forge packages to be maintained