@wesm 2018-09-28
Merging. The build only failing because of the known flakiness with the Java Flight tests and the new macOS failure that's shown up in the last 24 hours